protobuf proof of concept

+++ src/main/java/org/superbiz/game/msg/Part.java
@@ -1,11 +1,11 @@
 package org.superbiz.game.msg;
 
 public class Part {
-    private double x;
-    private double y;
-    private double r;
+    private float x;
+    private float y;
+    private float r;
 
-    public Part(double x, double y, double rotation) {
+    public Part(float x, float y, float rotation) {
         this.x = x;
         this.y = y;
         this.r = rotation;
@@ -14,27 +14,27 @@ public class Part {
     public Part() {
     }
 
-    public double getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(double x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public double getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(double y) {
+    public void setY(float y) {
         this.y = y;
     }
 
-    public double getR() {
+    public float getR() {
         return r;
     }
 
-    public void setR(double r) {
+    public void setR(float r) {
         this.r = r;
     }
 }
